Added support for limited bandwidth Config settings max download/upload speed

+++ mtTorrent/Api/FileTransfer.h
@@ -8,8 +8,8 @@ namespace mttApi
 	{
 	public:
 
-		API_EXPORT size_t getDownloadSpeed();
-		API_EXPORT size_t getUploadSpeed();
+		API_EXPORT uint32_t getDownloadSpeed();
+		API_EXPORT uint32_t getUploadSpeed();
 
 		API_EXPORT std::vector<mtt::ActivePeerInfo> getPeersInfo();
